refactor: more breaking out of files
